HIVE-25422
